# hackathon-max/openapi.yaml
openapi: 3.1.0
info:
  title: Профиль 10 — API
  description: Сервис подбора профиля обучения в 10 классе по направлениям подготовки, которые интересны
    ученику. Данные справочников в MVP демонстрационные.
  version: 1.0.0
paths:
...
  /api/classes/{class_code}:
    get:
      tags:
      - Справочники
      summary: Школа и профили по коду класса
      operationId: school_by_class_api_classes__class_code__get
      parameters:
      - name: class_code
        in: path
        required: true
        schema:
          type: string
          minLength: 3
          maxLength: 32
          title: Class Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response School By Class Api Classes  Class Code  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
servers:
- url: https://example.org
  description: Публичный адрес решения (заполняется при развёртывании)
- url: http://localhost:8080
  description: Локальный запуск через docker compose